moves get_raster_geo to a raster_utils.py
